chore: reorganization of public pages and indentation of the public routes file
