+++ src/pages/public/register/index.tsx
